Ended Classifier class with just one pass through the images.

+++ DetectMe/FileStorageHelper.h
@@ -17,7 +17,7 @@
 
 // For storing data
 
-+ (void)writeTemplate:(float *)vect withSize:(int *)size withTitle:(NSString *) filename;
++ (void)writeTemplate:(double *)vect withSize:(int *)size withTitle:(NSString *) filename;
 + (void)writeImage:(UInt8 *)vect withSize:(int *)size withTitle:(NSString *) filename;
 + (void)write:(double *)vect withSize:(int *)size withTitle:(NSString *) filename;
 + (void)writeImages:(UInt8 *)vect withSize:(int *)size withTitle:(NSString *) filename;
